Move synapse custom commands entirely to their own module
Typescripts namespare mixing works fine here
