+++ cypress/support/synapse.ts
@@ -20,14 +20,36 @@ import Chainable = Cypress.Chainable;
 import AUTWindow = Cypress.AUTWindow;
 import { SynapseInstance } from "../plugins/synapsedocker";
 
-export function startSynapse(template: string): Chainable<SynapseInstance> {
+declare global {
+    // eslint-disable-next-line @typescript-eslint/no-namespace
+    namespace Cypress {
+        interface Chainable {
+            /**
+             * Start a synapse instance with a given config template.
+             * @param template path to template within cypress/plugins/synapsedocker/template/ directory.
+             */
+            startSynapse(template: string): Chainable<SynapseInstance>;
+            /**
+             * Custom command wrapping task:synapseStop whilst preventing uncaught exceptions
+             * for if Synapse stopping races with the app's background sync loop.
+             * @param synapse the synapse instance returned by startSynapse
+             */
+            stopSynapse(synapse: SynapseInstance): Chainable<AUTWindow>;
+        }
+    }
+}
+
+function startSynapse(template: string): Chainable<SynapseInstance> {
     return cy.task<SynapseInstance>("synapseStart", template);
 }
 
-export function stopSynapse(synapse: SynapseInstance): Chainable<AUTWindow> {
+function stopSynapse(synapse: SynapseInstance): Chainable<AUTWindow> {
     // Navigate away from app to stop the background network requests which will race with Synapse shutting down
     return cy.window().then((win) => {
         win.location.href = 'about:blank';
         cy.task("synapseStop", synapse.synapseId);
     });
 }
+
+Cypress.Commands.add("startSynapse", startSynapse);
+Cypress.Commands.add("stopSynapse", stopSynapse);
